+++ le-ldc-collectiondb/src/main/java/com/latticeengines/ldc_collectiondb/entitymgr/impl/RawCollectionRequestMgrImpl.java
@@ -183,7 +183,13 @@ public class RawCollectionRequestMgrImpl extends JpaEntityMgrRepositoryImpl<RawC
     @Override
     public void cleanupTransferred(int batch) {
 
-        long minPid = readerRepository.getMinPid();
+        Long minPid = readerRepository.getMinPid();
+        if (minPid == null) {
+
+            return;
+
+        }
+
         repository.removeByTransferredAndPidLessThan(true, minPid + batch);
 
     }
